+++ lib/coub-worker.js
