+++ lib/emitter.js
