+++ lib/memory-cache.js
